requirements.txt
langgraph
langchain
langsmith
python-dotenv
langchain-groq 
langchain_tavily
langgraph-cli[inmem]
langchain-text-splitters
langchain-core
faiss-gpu
langchain-openai